Only get suggestions once when the input field is initially focused (#175)
Right now the suggestions are loaded whenever the field is empty
and it gets the focus. That's not needed, because old results are
retained anyway, thus we only ever need to get the suggestions
once. (If it gets refocused it displays the old suggestions again.)

This will save us quite some API requests and also will it save us
from race conditions when two such api requests are returning.

https://phabricator.wikimedia.org/T143645