Allow changing the weight given to classifying conditions
Allow changing the weight given to classifying conditions via
the $wgClassifyingConditionWeight setting. This only applies to
the "item" context if there is at least one classifying statement.

During testing it turned out that the classifying conditions
resulted in way better suggestions than the ones which are purely
based on whether a Statement with a given Property id exists.
Due to this, I want to be able to boost the weight of these
classifying conditions, in order to address T132839.

Note: This is only supported on MySQL and SQLite right now, like
the rest of this extension. On other DBMS, the setting will not
be taken into account.

https://phabricator.wikimedia.org/T132839